+++ src/libav.h
@@ -25,7 +25,7 @@
 
 #include <libavformat/avformat.h>
 #include <libavfilter/avfilter.h>
-#include "tvheadend.h"
+#include "esstream.h"
 
 /*
 Older versions of ffmpeg/libav don't have the AV_* prefix
