+++ src/packet.h
@@ -19,6 +19,8 @@
 #ifndef PACKET_H_
 #define PACKET_H_
 
+#include "esstream.h"
+
 struct memoryinfo;
 
 /**
