+++ src/transcoding/transcode.h
@@ -23,6 +23,7 @@
 
 
 #include "tvheadend.h"
+#include "streaming.h"
 
 
 /* TVHTranscoder ============================================================ */
