+++ src/tvheadend.h
@@ -120,16 +120,6 @@ typedef struct {
 
 
 /*
- * Commercial status
- */
-typedef enum {
-  COMMERCIAL_UNKNOWN,
-  COMMERCIAL_YES,
-  COMMERCIAL_NO,
-} th_commercial_advice_t;
-
-
-/*
  *
  */
 #define UILEVEL_DEFAULT  (-1)
@@ -262,7 +252,6 @@ TAILQ_HEAD(th_muxpkt_queue, th_muxpkt);
 LIST_HEAD(dvr_autorec_entry_list, dvr_autorec_entry);
 LIST_HEAD(dvr_timerec_entry_list, dvr_timerec_entry);
 TAILQ_HEAD(th_pktref_queue, th_pktref);
-LIST_HEAD(streaming_target_list, streaming_target);
 
 /**
  *
@@ -275,373 +264,6 @@ LIST_HEAD(streaming_target_list, streaming_target);
 #define TVH_KILL_USR2   5
 
 int tvh_kill_to_sig(int tvh_kill);
-
-/**
- * Stream component types
- */
-typedef enum {
-  SCT_NONE = -1,
-  SCT_UNKNOWN = 0,
-  SCT_RAW = 1,
-  SCT_PCR,     /* MPEG-TS PCR data */
-  SCT_CAT,     /* MPEG-TS CAT (EMM) data */
-  SCT_CA,      /* MPEG-TS ECM data */
-  SCT_HBBTV,   /* HBBTV info */
-  /* standard codecs */
-  SCT_MPEG2VIDEO,
-  SCT_MPEG2AUDIO,
-  SCT_H264,
-  SCT_AC3,
-  SCT_TELETEXT,
-  SCT_DVBSUB,
-  SCT_AAC,     /* AAC-LATM in MPEG-TS, ADTS + AAC in packet form */
-  SCT_MPEGTS,
-  SCT_TEXTSUB,
-  SCT_EAC3,
-  SCT_MP4A,    /* ADTS + AAC in MPEG-TS and packet form */
-  SCT_VP8,
-  SCT_VORBIS,
-  SCT_HEVC,
-  SCT_VP9,
-  SCT_THEORA,
-  SCT_OPUS,
-  SCT_FLAC,
-  SCT_LAST = SCT_FLAC
-} streaming_component_type_t;
-
-#define SCT_MASK(t) (1 << (t))
-
-#define SCT_ISVIDEO(t) ((t) == SCT_MPEG2VIDEO || (t) == SCT_H264 || \
-			(t) == SCT_VP8 || (t) == SCT_HEVC || \
-			(t) == SCT_VP9 || (t) == SCT_THEORA)
-
-#define SCT_ISAUDIO(t) ((t) == SCT_MPEG2AUDIO || (t) == SCT_AC3 || \
-			(t) == SCT_AAC  || (t) == SCT_MP4A || \
-			(t) == SCT_EAC3 || (t) == SCT_VORBIS || \
-			(t) == SCT_OPUS || (t) == SCT_FLAC)
-
-#define SCT_ISAV(t) (SCT_ISVIDEO(t) || SCT_ISAUDIO(t))
-
-#define SCT_ISSUBTITLE(t) ((t) == SCT_TEXTSUB || (t) == SCT_DVBSUB)
-
-/*
- * Scales for signal status values
- */
-typedef enum {
-  SIGNAL_STATUS_SCALE_UNKNOWN = 0,
-  SIGNAL_STATUS_SCALE_RELATIVE, // value is unsigned, where 0 means 0% and 65535 means 100%
-  SIGNAL_STATUS_SCALE_DECIBEL   // value is measured in dB * 1000
-} signal_status_scale_t;
-
-/**
- * The signal status of a tuner
- */
-typedef struct signal_status {
-  const char *status_text; /* adapter status text */
-  int snr;      /* signal/noise ratio */
-  signal_status_scale_t snr_scale;
-  int signal;   /* signal strength */
-  signal_status_scale_t signal_scale;
-  int ber;      /* bit error rate */
-  int unc;      /* uncorrected blocks */
-  int ec_bit;   /* error bit count */
-  int tc_bit;   /* total bit count */
-  int ec_block; /* error block count */
-  int tc_block; /* total block count */
-} signal_status_t;
-
-/**
- * Descramble info
- */
-typedef struct descramble_info {
-  uint16_t pid;
-  uint16_t caid;
-  uint32_t provid;
-  uint32_t ecmtime;
-  uint16_t hops;
-  char cardsystem[128];
-  char reader    [128];
-  char from      [128];
-  char protocol  [128];
-} descramble_info_t;
-
-/**
- *
- */
-typedef struct timeshift_status
-{
-  int     full;
-  int64_t shift;
-  int64_t pts_start;
-  int64_t pts_end;
-} timeshift_status_t;
-
-/**
- * Streaming skip
- */
-typedef struct streaming_skip
-{
-  enum {
-    SMT_SKIP_ERROR,
-    SMT_SKIP_REL_TIME,
-    SMT_SKIP_ABS_TIME,
-    SMT_SKIP_REL_SIZE,
-    SMT_SKIP_ABS_SIZE,
-    SMT_SKIP_LIVE
-  } type;
-  union {
-    off_t   size;
-    int64_t time;
-  };
-#if ENABLE_TIMESHIFT
-  timeshift_status_t timeshift;
-#endif
-} streaming_skip_t;
-
-
-/**
- * A streaming pad generates data.
- * It has one or more streaming targets attached to it.
- *
- * We support two different streaming target types:
- * One is callback driven and the other uses a queue + thread.
- *
- * Targets which already has a queueing intrastructure in place (such
- * as HTSP) does not need any interim queues so it would be a waste. That
- * is why we have the callback target.
- *
- */
-typedef struct streaming_pad {
-  struct streaming_target_list sp_targets;
-  int sp_ntargets;
-  int sp_reject_filter;
-} streaming_pad_t;
-
-
-TAILQ_HEAD(streaming_message_queue, streaming_message);
-
-/**
- * Streaming messages types
- */
-typedef enum {
-
-  /**
-   * Packet with data.
-   *
-   * sm_data points to a th_pkt. th_pkt will be unref'ed when
-   * the message is destroyed
-   */
-  SMT_PACKET,
-
-  /**
-   * Stream grace period
-   *
-   * sm_code contains number of seconds to settle things down
-   */
-
-  SMT_GRACE,
-
-  /**
-   * Stream start
-   *
-   * sm_data points to a stream_start struct.
-   * See transport_build_stream_start()
-   */
-
-  SMT_START,
-
-  /**
-   * Service status
-   *
-   * Notification about status of source, see TSS_ flags
-   */
-  SMT_SERVICE_STATUS,
-
-  /**
-   * Signal status
-   *
-   * Notification about frontend signal status
-   */
-  SMT_SIGNAL_STATUS,
-
-  /**
-   * Descrambler info message
-   *
-   * Notification about descrambler
-   */
-  SMT_DESCRAMBLE_INFO,
-
-  /**
-   * Streaming stop.
-   *
-   * End of streaming. If sm_code is 0 this was a result to an
-   * unsubscription. Otherwise the reason was external and the
-   * subscription scheduler will attempt to start a new streaming
-   * session.
-   */
-  SMT_STOP,
-
-  /**
-   * Streaming unable to start.
-   *
-   * sm_code indicates reason. Scheduler will try to restart
-   */
-  SMT_NOSTART,
-
-  /**
-   * Streaming unable to start (non-fatal).
-   *
-   * sm_code indicates reason. Scheduler will try to restart
-   */
-  SMT_NOSTART_WARN,
-
-  /**
-   * Raw MPEG TS data
-   */
-  SMT_MPEGTS,
-
-  /**
-   * Internal message to exit receiver
-   */
-  SMT_EXIT,
-
-  /**
-   * Set stream speed
-   */
-  SMT_SPEED,
-
-  /**
-   * Skip the stream
-   */
-  SMT_SKIP,
-
-  /**
-   * Timeshift status
-   */
-  SMT_TIMESHIFT_STATUS,
-
-} streaming_message_type_t;
-
-#define SMT_TO_MASK(x) (1 << ((unsigned int)x))
-
-
-#define SM_CODE_OK                        0
-
-#define SM_CODE_UNDEFINED_ERROR           1
-
-#define SM_CODE_FORCE_OK                  10
-
-#define SM_CODE_SOURCE_RECONFIGURED       100
-#define SM_CODE_BAD_SOURCE                101
-#define SM_CODE_SOURCE_DELETED            102
-#define SM_CODE_SUBSCRIPTION_OVERRIDDEN   103
-#define SM_CODE_INVALID_TARGET            104
-#define SM_CODE_USER_ACCESS               105
-#define SM_CODE_USER_LIMIT                106
-#define SM_CODE_WEAK_STREAM               107
-#define SM_CODE_USER_REQUEST              108
-#define SM_CODE_PREVIOUSLY_RECORDED       109
-
-#define SM_CODE_NO_FREE_ADAPTER           200
-#define SM_CODE_MUX_NOT_ENABLED           201
-#define SM_CODE_NOT_FREE                  202
-#define SM_CODE_TUNING_FAILED             203
-#define SM_CODE_SVC_NOT_ENABLED           204
-#define SM_CODE_BAD_SIGNAL                205
-#define SM_CODE_NO_SOURCE                 206
-#define SM_CODE_NO_SERVICE                207
-#define SM_CODE_NO_VALID_ADAPTER          208
-#define SM_CODE_NO_ADAPTERS               209
-#define SM_CODE_INVALID_SERVICE           210
-#define SM_CODE_CHN_NOT_ENABLED           211
-
-#define SM_CODE_ABORTED                   300
-
-#define SM_CODE_NO_DESCRAMBLER            400
-#define SM_CODE_NO_ACCESS                 401
-#define SM_CODE_NO_INPUT                  402
-#define SM_CODE_NO_SPACE                  403
-
-typedef enum
-{
-  SIGNAL_UNKNOWN,
-  SIGNAL_GOOD,
-  SIGNAL_BAD,
-  SIGNAL_FAINT,
-  SIGNAL_NONE
-} signal_state_t;
-
-static struct strtab signal_statetab[] = {
-  { "GOOD",       SIGNAL_GOOD    },
-  { "BAD",        SIGNAL_BAD     },
-  { "FAINT",      SIGNAL_FAINT   },
-  { "NONE",       SIGNAL_NONE    },
-};
-
-static inline const char * signal2str ( signal_state_t st )
-{
-  const char *r = val2str(st, signal_statetab);
-  if (!r) r = "UNKNOWN";
-  return r;
-}
-
-/**
- * Streaming messages are sent from the pad to its receivers
- */
-typedef struct streaming_message {
-  TAILQ_ENTRY(streaming_message) sm_link;
-  streaming_message_type_t sm_type;
-#if ENABLE_TIMESHIFT
-  int64_t sm_time;
-#endif
-  union {
-    void *sm_data;
-    int sm_code;
-  };
-} streaming_message_t;
-
-/**
- * A streaming target receives data.
- */
-
-typedef void (st_callback_t)(void *opaque, streaming_message_t *sm);
-
-typedef struct {
-  st_callback_t *st_cb;
-  htsmsg_t *(*st_info)(void *opaque, htsmsg_t *list);
-} streaming_ops_t;
-
-typedef struct streaming_target {
-  LIST_ENTRY(streaming_target) st_link;
-  streaming_pad_t *st_pad;               /* Source we are linked to */
-  streaming_ops_t st_ops;
-  void *st_opaque;
-  int st_reject_filter;
-} streaming_target_t;
-
-
-/**
- *
- */
-typedef struct streaming_queue {
-
-  streaming_target_t sq_st;
-
-  pthread_mutex_t sq_mutex;    /* Protects sp_queue */
-  tvh_cond_t  sq_cond;         /* Condvar for signalling new packets */
-
-  size_t          sq_maxsize;  /* Max queue size (bytes) */
-  size_t          sq_size;     /* Actual queue size (bytes) - only data */
-
-  struct streaming_message_queue sq_queue;
-
-} streaming_queue_t;
-
-
-streaming_component_type_t streaming_component_txt2type(const char *str);
-const char *streaming_component_type2txt(streaming_component_type_t s);
-streaming_component_type_t streaming_component_txt2type(const char *s);
-const char *streaming_component_audio_type2desc(int audio_type);
 
 static inline unsigned int tvh_strhash(const char *s, unsigned int mod)
 {
@@ -665,8 +287,6 @@ int tvh_str_update(char **strp, const char *src);
 
 int sri_to_rate(int sri);
 int rate_to_sri(int rate);
-
-extern struct service_list all_transports;
 
 extern void scopedunlock(pthread_mutex_t **mtxp);
